gpu: pvr: Enable runtime PM support for ddk 1.17
This is what works with the old pvr-omap4-dkms nicely for idling the
clkctrl clock for sgx when not in use.

Looks like 1.17 ddk blobs are built without config option for
SUPPORT_ACTIVE_POWER_MANAGEMENT though. So we cannot enable the
option for SUPPORT_ACTIVE_POWER_MANAGEMENT in Makefile yet.
Otherwise we get a config mismatch on initializing the blobs.

But hopefully eventually we can just have the option enabled for
SUPPORT_ACTIVE_POWER_MANAGEMENT for all the SoCs and blobs.

Signed-off-by: Tony Lindgren <tony@atomide.com>